refactor: ReportRepository에서 사용하는 DataSource 별도 클래스로 분리
