+++ data/src/main/java/com/bestapp/zipbab/data/datasource/remote/ReportRemoteDataSource.kt
@@ -1,6 +1,6 @@
-package com.bestapp.zipbab.data.repository
+package com.bestapp.zipbab.data.datasource.remote
 
-interface ReportRepository {
+interface ReportRemoteDataSource {
 
     suspend fun reportUser(userDocumentID: String): Boolean
     suspend fun reportPost(userDocumentID: String, postDocumentID: String): Boolean
